+++ src/Common/JK.Common.Tests/ComplexObject.cs
@@ -3,7 +3,7 @@ using System;
 
 namespace JK.Common.Tests
 {
-    public class ComplexObject : IIdentifiable<object>
+    public class ComplexObject : IIdentifiable<int>
     {
         public ComplexObject()
         {
@@ -16,6 +16,11 @@ namespace JK.Common.Tests
             this.FloatProperty = 0;
             this.BooleanProperty = false;
             this.ChildObjectSimple = new SimpleObject();
+        }
+
+        public ComplexObject(DateTime dateTime) : this()
+        {
+            this.DateTimeProperty = dateTime;
         }
 
         public string StringProperty { get; set; }
@@ -36,6 +41,6 @@ namespace JK.Common.Tests
 
         public ComplexObject ChildObjectComplex { get; set; }
 
-        public object Id { get; set; }
+        public int Id { get; set; }
     }
 }